requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
black==24.10.0
    # via -r requirements/requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via black
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
flatbuffers==24.3.25
    # via onnxruntime
fonttools==4.55.0
    # via matplotlib
fsspec==2024.10.0
    # via
    #   huggingface-hub
    #   torch
ftfy==6.3.1
    # via open-clip-torch
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via openai
huggingface-hub==0.26.2
    # via
    #   -r requirements/requirements.in
    #   open-clip-torch
    #   timm
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.4.5
    # via matplotlib
jinja2==3.1.4
    # via torch
jiter==0.7.1
    # via openai
kiwisolver==1.4.7
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.2
    # via
    #   -r requirements/requirements.in
    #   pycocotools
mpmath==1.3.0
    # via sympy
mypy-extensions==1.0.0
    # via black
networkx==3.2.1
    # via torch
numpy==2.0.2
    # via
    #   -r requirements/requirements.in
    #   contourpy
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pycocotools
    #   torchvision
onnx==1.17.0
    # via -r requirements/requirements.in
onnxruntime==1.19.2
    # via -r requirements/requirements.in
open-clip-torch==2.29.0
    # via -r requirements/requirements.in
openai==1.54.4
    # via -r requirements/requirements.in
opencv-python==4.10.0.84
    # via -r requirements/requirements.in
packaging==24.2
    # via
    #   black
    #   huggingface-hub
    #   matplotlib
    #   onnxruntime
pathspec==0.12.1
    # via black
pillow==11.0.0
    # via
    #   matplotlib
    #   torchvision
platformdirs==4.3.6
    # via black
protobuf==5.28.3
    # via
    #   onnx
    #   onnxruntime
pycocotools==2.0.8
    # via -r requirements/requirements.in
pydantic==2.9.2
    # via openai
pydantic-core==2.23.4
    # via pydantic
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   timm
regex==2024.11.6
    # via open-clip-torch
requests==2.32.3
    # via huggingface-hub
safetensors==0.4.5
    # via
    #   open-clip-torch
    #   timm
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sympy==1.13.1
    # via
    #   onnxruntime
    #   torch
timm==1.0.11
    # via open-clip-torch
tomli==2.1.0
    # via black
torch==2.5.1
    # via
    #   open-clip-torch
    #   timm
    #   torchvision
torchvision==0.20.1
    # via
    #   open-clip-torch
    #   timm
tqdm==4.67.0
    # via
    #   huggingface-hub
    #   open-clip-torch
    #   openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   black
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.2.3
    # via requests
wcwidth==0.2.13
    # via ftfy
zipp==3.21.0
    # via importlib-resources
